Added import of offspring. This significantly increased the number of operations needed in order to import pedigree data from dogsearch, so the import concurrency has been improved.

+++ src/main/java/no/nkk/dogpopulation/graph/dogbuilder/InLitterRelationshipBuilder.java
@@ -1,5 +1,6 @@
 package no.nkk.dogpopulation.graph.dogbuilder;
 
+import no.nkk.dogpopulation.graph.Builder;
 import no.nkk.dogpopulation.graph.DogGraphConstants;
 import no.nkk.dogpopulation.graph.DogGraphRelationshipType;
 import org.neo4j.graphdb.Direction;
@@ -16,8 +17,8 @@ public class InLitterRelationshipBuilder extends AbstractRelationshipBuilder {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(InLitterRelationshipBuilder.class);
 
-    private LitterNodeBuilder litterBuilder;
-    private DogNodeBuilder puppyBuilder;
+    private Builder<Node> litterBuilder;
+    private Builder<Node> puppyBuilder;
     private Node litter;
     private Node puppy;
 
@@ -37,15 +38,30 @@ public class InLitterRelationshipBuilder extends AbstractRelationshipBuilder {
             if (existingLitter.equals(this.litter)) {
                 return existingInLitter; // already connected to correct litter
             }
+            String litterId = (String) this.litter.getProperty(DogGraphConstants.LITTER_ID);
+            if (litterId.trim().isEmpty()) {
+                return existingInLitter;
+            }
             String puppyUuid = (String) puppy.getProperty(DogGraphConstants.DOG_UUID);
             String existingLitterId = (String) existingLitter.getProperty(DogGraphConstants.LITTER_ID);
-            String litterId = (String) this.litter.getProperty(DogGraphConstants.LITTER_ID);
             LOGGER.warn("LITTER CONFLICT: Dog {} is already in litter {} but will now be moved to litter {}.", puppyUuid, existingLitterId, litterId);
             existingInLitter.delete();
         }
         return puppy.createRelationshipTo(this.litter, DogGraphRelationshipType.IN_LITTER);
     }
 
+    @Override
+    public void reset() {
+        if (litterBuilder != null) {
+            litterBuilder.reset();
+            litter = null;
+        }
+        if (puppyBuilder != null) {
+            puppyBuilder.reset();
+            puppy = null;
+        }
+        super.reset();
+    }
 
     public InLitterRelationshipBuilder litter(Node litter) {
         this.litter = litter;
@@ -55,11 +71,11 @@ public class InLitterRelationshipBuilder extends AbstractRelationshipBuilder {
         this.puppy = puppy;
         return this;
     }
-    public InLitterRelationshipBuilder litter(LitterNodeBuilder litterBuilder) {
+    public InLitterRelationshipBuilder litter(Builder<Node> litterBuilder) {
         this.litterBuilder = litterBuilder;
         return this;
     }
-    public InLitterRelationshipBuilder puppy(DogNodeBuilder puppyBuilder) {
+    public InLitterRelationshipBuilder puppy(Builder<Node> puppyBuilder) {
         this.puppyBuilder = puppyBuilder;
         return this;
     }
